Move the details of C99 identifier mangling out of `ProjectModel` and down into `Utility`. It looks as if at some point this was meant to be able to fail because it might remove all the characters and end up with an empty string, but now that it replaces characters it doesn't seem possible for it to end up with an empty string assuming it was given a non-empty string. Instead we just define it to never generate an empty string if given a non-empty string. It will continue to be the caller's responsibility to check separately for non-empty strings, for which it is likely to want different errors anyway (e.g. reporting a missing module name as different from a malformed module name or whatever).
