+++ Tests/PackageModelTests/XCTestManifests.swift
@@ -13,7 +13,6 @@ import XCTest
 #if !os(macOS)
 public func allTests() -> [XCTestCaseEntry] {
     return [
-        testCase(c99nameTests.allTests),
         testCase(ModuleTests.allTests),
     ]
 }
